+++ test/test_indexed_algebra.py
@@ -2,7 +2,15 @@ import unittest
 
 import sympy as sp
 
-from pyopl._indexed_algebra import DecisionAtom, IndexTerm, QuantifiedExpression, render_indexed_ir
+from pyopl._indexed_algebra import (
+    DecisionAtom,
+    DomainTerm,
+    IndexTerm,
+    QuantifiedExpression,
+    domain_dependency_graph,
+    free_binders,
+    render_indexed_ir,
+)
 
 
 class IndexedAlgebraRenderingTests(unittest.TestCase):
@@ -21,6 +29,36 @@ class IndexedAlgebraRenderingTests(unittest.TestCase):
         atom = DecisionAtom("x", (IndexTerm("binder", 0), index))
 
         self.assertEqual(render_indexed_ir(atom), "x[$0, ($1 + 1)]")
+
+    def test_free_binders_traverses_nested_index_tree(self):
+        index = IndexTerm("arithmetic", ("+", IndexTerm("binder", 1), IndexTerm("number", "1")))
+        atom = DecisionAtom("x", (IndexTerm("binder", 0), index))
+
+        self.assertEqual(free_binders(atom), frozenset({0, 1}))
+
+    def test_render_parameter_selected_index(self):
+        selected = IndexTerm("application", ("next", (IndexTerm("binder", 0),)))
+
+        self.assertEqual(render_indexed_ir(selected), "next[$0]")
+
+    def test_domain_dependency_graph_tracks_prior_binders(self):
+        domains = (
+            DomainTerm("named", "I"),
+            DomainTerm("range", (IndexTerm("number", "1"), IndexTerm("binder", 0))),
+        )
+
+        self.assertEqual(domain_dependency_graph(domains, first_binder_id=0), {0: frozenset(), 1: frozenset({0})})
+
+    def test_quantifier_free_binders_excludes_local_scope(self):
+        outer = IndexTerm("binder", 4)
+        local = IndexTerm("binder", 5)
+        expression = QuantifiedExpression(
+            (DomainTerm("range", (IndexTerm("number", "1"), outer)),),
+            ("compare", ">=", ("binder", 5), ("number", "1")),
+            (sp.S.Zero, ((DecisionAtom("x", (local, outer)), sp.S.One),)),
+        )
+
+        self.assertEqual(free_binders(expression, first_binder_id=5), frozenset({4}))
 
 
 if __name__ == "__main__":